Security review (#130–#154) + feature round (#120–#129): 23 fixes
Two rounds since the initial hardening publication.

Security — a whole-codebase multi-model review of the core-critical subset
(the four CPU instruction cores, the dynamic-translation engine, the guest→host
memory boundary, the file loaders, network, disk, debugger, and the highest-risk
devices): parallel Claude review agents, cross-checked by three independent
cloud models (GLM / DeepSeek-V3 / Qwen3-Coder), then adjudicated against the
source. ~23 confirmed fixes, every one pre-existing in the baseline:

- CRITICAL #137 cpus/cpu_mips_instr.c memset combiner — an unsigned end-start
  underflow that ALSO wrapped the page-boundary clamp drove a multi-gigabyte
  memset straight into the host heap (guest-triggerable on pmax). Fall back to
  the safe per-instruction path when end < start.
- HIGH #145 devices/dev_pvr.c — oversized guest display geometry overran the
  fixed host framebuffer (guest→host heap write). Clamp to the host framebuffer.
- HIGH #149 file/file_srec.c — an over-long S-record count over-read ~4 KB of
  host stack into guest RAM. Clamp to the parsed record length.
- plus a medium/low tail: guest-set SCSI transfer-size overflow (OOM-exit DoS),
  a TCP timestamp-option over-read, a PPC time-base that never incremented,
  recoverable dev_osiop exit(1)s converted to warn-and-continue, free() on an
  mmap-backed cpu, and format-string / odd-length-checksum / NULL-deref /
  delay-slot / FDIV-by-zero / CLI-argument / signed-shift-UB hardening.

The dynamic-translation engine, the memory boundary, and the ELF/ECOFF/Mach-O/
a.out loaders were re-confirmed sound (no OOB).

Features (#120–#129): SuperH unaligned-access exceptions + 64-bit fmov
alignment; multi-track CUE/BIN CD images; testmips RAM above 256 MB;
subsystem-level debug breakpoints; debugger conveniences.

Regression-gated: a 9-machine multi-architecture boot sweep plus the full
OpenBSD/pmax rig (boot, root shell, NAT, disk, clean halt); build 0 errors /
0 warnings. Per-fix detail in CHANGELOG.md / REVIEW_FINDINGS.md; the full
source diff is refreshed in CHANGES.patch.

+++ src/cpus/cpu_arm_instr.c
@@ -1266,7 +1266,7 @@ X(swp)
 		fatal("swp: load failed\n");
 		return;
 	}
-	data = d[0] + (d[1] << 8) + (d[2] << 16) + (d[3] << 24);
+	data = d[0] + (d[1] << 8) + (d[2] << 16) + ((uint32_t)d[3] << 24);
 	data2 = reg(ic->arg[1]);
 	d[0] = data2; d[1] = data2 >> 8; d[2] = data2 >> 16; d[3] = data2 >> 24;
 	if (!cpu->memory_rw(cpu, cpu->mem, addr, d, sizeof(d), MEM_WRITE,
@@ -1409,11 +1409,11 @@ void arm_pop(struct cpu* cpu, uint32_t* np, int p_bit, int u_bit, int s_bit, int
 			if (cpu->byte_order == EMUL_LITTLE_ENDIAN) {
 				value = data[0] +
 				    (data[1] << 8) + (data[2] << 16)
-				    + (data[3] << 24);
+				    + ((uint32_t)data[3] << 24);
 			} else {
 				value = data[3] +
 				    (data[2] << 8) + (data[1] << 16)
-				    + (data[0] << 24);
+				    + ((uint32_t)data[0] << 24);
 			}
 		}
 
@@ -1717,10 +1717,10 @@ X(ldrex)
 
 	if (cpu->byte_order == EMUL_LITTLE_ENDIAN)
 		reg(ic->arg[0]) = word[0] + (word[1] << 8)
-		    + (word[2] << 16) + (word[3] << 24);
+		    + (word[2] << 16) + ((uint32_t)word[3] << 24);
 	else
 		reg(ic->arg[0]) = word[3] + (word[2] << 8)
-		    + (word[1] << 16) + (word[0] << 24);
+		    + (word[1] << 16) + ((uint32_t)word[0] << 24);
 }
 Y(ldrex)
 /*
